ci: reorganize the global tests
