+++ test/global/integration/connection/test-change-user.test.mts
@@ -1,8 +1,8 @@
-import type { RowDataPacket } from '../../index.js';
+import type { RowDataPacket } from '../../../../index.js';
 import { Buffer } from 'node:buffer';
 import process from 'node:process';
 import { describe, it, skip, strict } from 'poku';
-import { createConnection } from '../common.test.mjs';
+import { createConnection } from '../../../common.test.mjs';
 
 if (`${process.env.MYSQL_CONNECTION_URL}`.includes('pscale_pw_')) {
   skip('Skipping test for PlanetScale');
